OTP router instance moved into object for sharing.
